DIVIDE :: MainScreen
