+++ lib/Screen/MainPage.dart
@@ -1,0 +1,50 @@
+import 'package:flutter/material.dart';
+import 'package:quote_app/logic/GetQuoteData.dart';
+
+class MainPage extends StatefulWidget {
+  const MainPage({super.key, required this.title});
+
+  final String title;
+
+  @override
+  State<MainPage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MainPage> {
+  int _counter = 0;
+
+  void _incrementCounter() {
+    setState(() {
+      _counter++;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text(
+              'Push Button To Quote.',
+            ),
+            Text(
+              '$_counter',
+              style: Theme.of(context).textTheme.headlineMedium,
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: fetchQuotes,
+        tooltip: 'Increment',
+        child: const Icon(Icons.add),
+      ),
+    );
+  }
+}
